refactorizacion del signup seseparo las funciones del firebase y las del dom
