+++ package.json
@@ -32,7 +32,7 @@
     "firebase-mock": "^2.3.2",
     "gh-pages": "^3.2.3",
     "htmlhint": "^0.14.0",
-    "jest": "^27.0.6",
+    "jest": "^27.1.0",
     "regenerator-runtime": "^0.13.1",
     "serve": "^12.0.0",
     "stylelint": "^13.13.1",
